pyproject.toml
[tool.poetry]
name = "yatlogger"
version = "0.1.0"
description = "Yet another telegram logger"
authors = ["cyd3r <cyd3rhacker@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
